Ecommerce_MyStore/src/test/java/com/myStore/testCases/AddToCartPageTest.java
package com.myStore.testCases;

import org.junit.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Parameters;
import org.testng.annotations.Test;

import com.mystore.base.BaseClass;
import com.mystore.pageobjects.AddToCartPage;
import com.mystore.pageobjects.IndexPage;
import com.mystore.pageobjects.SearchResultPage;

public class AddToCartPageTest extends BaseClass {
	IndexPage ip;
	SearchResultPage searchresult;
	AddToCartPage addToCartPage;
	
	@Parameters("browser")
	@BeforeMethod(groups= {"Smoke","Sanity","Regression"})
	public void setup(String browser)
	{
		launchApp(browser);
	}

	@AfterMethod(groups= {"Smoke","Sanity","Regression"})
	public void tearDown() {
		getDriver().quit();
	}

	@Test(groups= {"Regression","Sanity"})
	public void addToCartTest() {
		ip=new IndexPage();
		searchresult=ip.searchProduct("dress");
		addToCartPage=searchresult.clickOnProduct();
		addToCartPage.enterQuantity("2");
		addToCartPage.selectSize("M");
		addToCartPage.clickOnAddToCart();
		boolean RESULT=addToCartPage.validateAddToCart();
		Assert.assertTrue(RESULT);
		
	}

}
